+++ SRC/chpevd.f
@@ -300,7 +300,7 @@
      $   RETURN
 *
       IF( N.EQ.1 ) THEN
-         W( 1 ) = AP( 1 )
+         W( 1 ) = REAL( AP( 1 ) )
          IF( WANTZ )
      $      Z( 1, 1 ) = CONE
          RETURN
